search state and transition

+++ app/src/main/java/com/sovathna/khmerdictionary/domain/model/intent/WordListIntent.kt
@@ -8,5 +8,5 @@ sealed class WordListIntent : MviIntent {
     val offset: Int
   ) : WordListIntent()
 
-  data class Select(val current: Int?) : WordListIntent()
+  data class Select(val current: Long?) : WordListIntent()
 }
